+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 # [1.6.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.6.0) (xx-06-2017)
 ## Fix
 - [Move /task-forms/{task-id}/variables from rest to enterprise](https://github.com/Alfresco/alfresco-js-api/issues/199)
+- [Added query param for get related contents](https://issues.alfresco.com/jira/browse/ADF-684)
 
 <a name="1.5.0"></a>
 # [1.5.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.5.0) (25-05-2017)