+++ CHANGELOG.md
@@ -6,14 +6,13 @@
 
 <a name="1.6.0"></a>
 # [1.6.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.6.0) (xx-06-2017)
-## Features
-- [Get rendition URL for a given nodeId from content API](https://github.com/Alfresco/alfresco-js-api/issues/237)
 ## Fix
 - [Move /task-forms/{task-id}/variables from rest to enterprise](https://github.com/Alfresco/alfresco-js-api/issues/199)
 - [Added query param for get related contents](https://issues.alfresco.com/jira/browse/ADF-684)
 
 ## Features
 - [Complete TypeScript definitions for JS-API](https://github.com/Alfresco/alfresco-js-api/issues/228)
+- [Get rendition URL for a given nodeId from content API](https://github.com/Alfresco/alfresco-js-api/issues/237)
 
 <a name="1.5.0"></a>
 # [1.5.0](https://github.com/Alfresco/alfresco-js-api/releases/tag/1.5.0) (25-05-2017)